Abort submitted job

Abort submitted job

stopWorkflow(workspaceName, submissionId = NULL, dry = TRUE)

Arguments

workspaceName

Name of the workspace

submissionId

A character. Submission ID you want to abort. You can find the submission id using monitorSubmission function. If it is not defined, the most recent submission will be aborted.

dry

Logical(1) when TRUE (default), report the consequences but do not perform the action requested. When FALSE, perform the action.

Value

This function will print out whether the call for workflow abortion was successful or not. In case it was unsuccesful, the diagnosis will be suggested as a part of the message.

Examples

library(AnVIL)
if (gcloud_exists() && nzchar(avworkspace_name())) {
if ("salmon" %in% avworkspaces()$name)
stopWorkflow(workspaceName = "salmon")
}